+++ httpx_auth/authentication.py
@@ -3,7 +3,7 @@ import os
 import uuid
 from hashlib import sha256, sha512
 from urllib.parse import parse_qs, urlsplit, urlunsplit, urlencode
-from typing import Optional, Generator
+from typing import Optional, Generator, Union, Iterable
 
 import httpx
 
@@ -59,6 +59,17 @@ def _get_query_parameter(url: str, param_name: str) -> Optional[str]:
     return all_values[0] if all_values else None
 
 
+def _content_from_response(response: httpx.Response) -> dict:
+    content_type = response.headers.get("content-type")
+    if content_type == "text/html; charset=utf-8":
+        return {
+            key_values[0]: key_values[1]
+            for key_value in response.text.split("&")
+            if (key_values := key_value.split("=")) and len(key_values) == 2
+        }
+    return response.json()
+
+
 def request_new_grant_with_post(
     url: str, data, grant_name: str, client: httpx.Client
 ) -> (str, int):
@@ -68,7 +79,7 @@ def request_new_grant_with_post(
         # As described in https://tools.ietf.org/html/rfc6749#section-5.2
         raise InvalidGrantRequest(response)
 
-    content = response.json()
+    content = _content_from_response(response)
     token = content.get(grant_name)
     if not token:
         raise GrantNotProvided(grant_name, content)
@@ -1047,6 +1058,66 @@ class OktaAuthorizationCode(OAuth2AuthorizationCode):
             f"https://{instance}/oauth2/{authorization_server}/v1/authorize",
             f"https://{instance}/oauth2/{authorization_server}/v1/token",
             client_id=client_id,
+            **kwargs,
+        )
+
+
+class WakaTimeAuthorizationCode(OAuth2AuthorizationCode):
+    """
+    Describes a WakaTime (OAuth 2) "Access Token" authorization code flow requests authentication.
+    """
+
+    def __init__(
+        self,
+        client_id: str,
+        client_secret: str,
+        scope: Union[str, Iterable[str]],
+        **kwargs,
+    ):
+        """
+        :param client_id: WakaTime Application Identifier (formatted as an Universal Unique Identifier)
+        :param client_secret: WakaTime Application Secret (formatted as waka_sec_ followed by an Universal Unique Identifier)
+        :param scope: Scope parameter sent in query. Can also be a list of scopes.
+        :param response_type: Value of the response_type query parameter.
+        token by default.
+        :param token_field_name: Name of the expected field containing the token.
+        access_token by default.
+        :param early_expiry: Number of seconds before actual token expiry where token will be considered as expired.
+        Default to 30 seconds to ensure token will not expire between the time of retrieval and the time the request
+        reaches the actual server. Set it to 0 to deactivate this feature and use the same token until actual expiry.
+        :param nonce: Refer to http://openid.net/specs/openid-connect-core-1_0.html#IDToken for more details
+        (formatted as an Universal Unique Identifier - UUID). Use a newly generated UUID by default.
+        :param redirect_uri_endpoint: Custom endpoint that will be used as redirect_uri the following way:
+        http://localhost:<redirect_uri_port>/<redirect_uri_endpoint>. Default value is to redirect on / (root).
+        :param redirect_uri_port: The port on which the server listening for the OAuth 2 token will be started.
+        Listen on port 5000 by default.
+        :param timeout: Maximum amount of seconds to wait for a token to be received once requested.
+        Wait for 1 minute by default.
+        :param success_display_time: In case a token is successfully received,
+        this is the maximum amount of milliseconds the success page will be displayed in your browser.
+        Display the page for 1 millisecond by default.
+        :param failure_display_time: In case received token is not valid,
+        this is the maximum amount of milliseconds the failure page will be displayed in your browser.
+        Display the page for 5 seconds by default.
+        :param header_name: Name of the header field used to send token.
+        Token will be sent in Authorization header field by default.
+        :param header_value: Format used to send the token value.
+        "{token}" must be present as it will be replaced by the actual token.
+        Token will be sent as "Bearer {token}" by default.
+        :param client: httpx.Client instance that will be used to request the token.
+        Use it to provide a custom proxying rule for instance.
+        :param kwargs: all additional authorization parameters that should be put as query parameter
+        in the authorization URL.
+        """
+        if not scope:
+            raise Exception("Scope is mandatory.")
+        OAuth2AuthorizationCode.__init__(
+            self,
+            "https://wakatime.com/oauth/authorize",
+            "https://wakatime.com/oauth/token",
+            client_id=client_id,
+            client_secret=client_secret,
+            scope=",".join(scope) if isinstance(scope, list) else scope,
             **kwargs,
         )